adding migrations

+++ contracts/Funder.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.4.22<0.9.0;
 
 contract Funder{
-    uint public numOfFunders;
+    uint256 public numOfFunders;
     mapping(uint=>address) private funders;
 
     receive() external payable {}
@@ -11,7 +11,7 @@ contract Funder{
         funders[numOfFunders] = msg.sender;
     }
 
-    function withdraw(uint withdrawAmount) external {
+    function withdraw(uint256 withdrawAmount) external {
         require(withdrawAmount <= 2000000000000000000, "Cannot withdraw more than 2 ether");
         payable(msg.sender).transfer(withdrawAmount);
     }
